requirements.txt
Pillow
openpyxl
numpy
scikit-learn
scipy
